functions/mark_trips/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
